pc: Add support for virtio-iommu-pci
The virtio-iommu-pci is instantiated through the -device QEMU
option. However if instantiated it also requires an IORT ACPI table
to describe the ID mappings between the root complex and the iommu.

This patch adds the generation of the IORT table if the
virtio-iommu-pci device is instantiated.

We also declare the [0xfee00000 - 0xfeefffff] MSI reserved region
so that it gets bypassed by the IOMMU.

Signed-off-by: Eric Auger <eric.auger@redhat.com>