hw/rtc/pl031: Send RTC_CHANGE QMP event
The PL031 currently is not able to report guest RTC change to the QMP
monitor as opposed to mc146818 or spapr RTCs. This patch adds the call
to qapi_event_send_rtc_change() when the Load Register is written. The
value that is reported corresponds to the difference between the new
RTC value and the RTC value elapsed since the base.

For instance adding 20s to the guest RTC value will report 20:
{'timestamp': {'seconds': 1631189494, 'microseconds': 16932},
'event': 'RTC_CHANGE', 'data': {'offset': 20}}

Adding another extra 20s to the guest RTC value will report 40:
{'timestamp': {'seconds': 1631189498, 'microseconds': 9708},
'event': 'RTC_CHANGE', 'data': {'offset': 40}}

To compute the offset we need to track the origin tick_offset (the one
computed at init time). So we need to migrate that field, which is done
in a dedicated subsection. The migration of this subsection is disabled
for machine types less or equal than 6.1.

After migration, adding an extra 20s on the destination returns 60:
{'timestamp': {'seconds': 1631189522, 'microseconds': 13081},
'event': 'RTC_CHANGE', 'data': {'offset': 60}}

Signed-off-by: Eric Auger <eric.auger@redhat.com>

---

Tested with the following script ran on guest:

  #!/bin/sh
  old=$(hwclock --show | cut -f1-7 -d' ')
  oldabs=$(date +%s -d "$old")
  newabs=$(expr $oldabs + $1)
  new=$(date -d @"$newabs")
  echo Old: $oldabs $old
  echo New: $newabs $new
  hwclock --set --date "$new"

This was tested with both -rtc base=2010-12-03T01:02:00 and base=utc
qemu options. As far as I can see the reported value match what
is observed on x86 (except on x86 the values are not exactly the one
used on guest, ie. 18 for instance instead of 20).

Without migrating the original tick_offset (ie. original_tick_offset
taking the destination tick_offset value), a delta is observed. I
don't know whether it is normal or a bug. At firt glance I had the
impression it worked without the migration scheme but this delta
urged me to migrate the original_offset too.