hw/vfio/common: Silence ram device offset alignment error traces
Failing to DMA MAP a ram_device should not cause an error message.
This is currently happening with the TPM CRB command region and
this is causing confusion.

We may want to keep the trace for debug purpose though.

Signed-off-by: Eric Auger <eric.auger@redhat.com>

---

I am not totally clear why we do not fail on the non RAM device case
though.