vfio/common: Split common.c into common.c and group.c
Before introducing the support for /dev/iommu in VFIO let's try
to separate code that is not going to be impacted from the one
bound to be touched. Not impacted code is kept in common.c while
the rest (using containers, groups, ...) is moved to group.c for
the time being.

Few functions become public:
- vfio_get_xlat_addr
- vfio_kvm_device_add/del_group
- vfio_get_iommu_info_cap

also vfio_kvm_device_fd becomes public

No code change intended

Signed-off-by: Yi Liu <yi.l.liu@intel.com>
Signed-off-by: Eric Auger <eric.auger@redhat.com>