hw/vfio/pci: Add an iommufd option to select the iommu backend
Now we support two iommu backends, the legacy vfio iommu one and
iommufd, let's expose an iommufd option that allow the user to
select the backend. In auto mode, it chooses the iommufd backend
if this latter is exposed by the host.

For other devices which are not yet exposed with such option,
the first working backend still is used.

Signed-off-by: Eric Auger <eric.auger@redhat.com>