hw/vfio/pci: Add an iommufd option to select the iommu backend
Now we support two iommu backends, the legacy vfio iommu one and
iommufd, let's expose an iommufd option that allow the user to
select the backend. In auto mode, it chooses the iommufd backend
if this latter is exposed bu the host.

Signed-off-by: Eric Auger <eric.auger@redhat.com>

---

it is arguable if we shall use the iommufd ops in auto mode and
by default. Maybe we shall do the oppositive and be conservative