vfio: Add base object for VFIOContainer
This new object is a base object for derived VFIOContainer and iommufd
"container". It has the common fields used by both while the derived
objects will specialize the object for a given backend.

Below is the object. It's named as VFIOContainer, old VFIOContainer is
replaced with VFIOLegacyContainer.

struct VFIOContainer {
    Object parent_obj;

    VFIOAddressSpace *space;
    MemoryListener listener;
    Error *error;
    bool initialized;
    bool dirty_pages_supported;
    uint64_t dirty_pgsizes;
    uint64_t max_dirty_bitmap_size;
    unsigned long pgsizes;
    unsigned int dma_max_mappings;
    QLIST_HEAD(, VFIOGuestIOMMU) giommu_list;
    QLIST_HEAD(, VFIOHostDMAWindow) hostwin_list;
    QLIST_HEAD(, VFIORamDiscardListener) vrdl_list;
    QLIST_ENTRY(VFIOBaseContainer) next;
};

TODO:
The fields of the base object (like the pgsizes) is initialized in the
BE. To be an independent module, such works may be better done in the
object file.

Co-authored-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Eric Auger <eric.auger@redhat.com>
Signed-off-by: Yi Liu <yi.l.liu@intel.com>