vfio: Add vfio_get/put_device()
vfio_get_device() - attaches the device to an IOMMU backend.
The function picks up the first backend for which the attachment
is successful.  Once the function returns, the caller can use
the device fd to access the device.

vfio_put_device() - release the device. The secure context will
be destroyed, device fd will be closed. No more access to the
device is allowed.

Signed-off-by: Yi Liu <yi.l.liu@intel.com>