cover-letter: vfio: Adopt iommufd
With the introduction of iommufd[1], linux kernel provides a generic
interface for userspace drivers to propagate their DMA mappings to kernel
for assigned devices. This series does the porting of the VFIO devices
onto the /dev/iommu uapi and let it coexist with the legacy implementation.
Other devices like vpda, vfio mdev is not considered yet.

For vfio devices, the new interface is tied with device fd and iommufd
as iommufd solution is device-centric. This is different from legacy
vfio which is group-centric. To support both interfaces in QEMU, this
series introduces the iommu backend concept in the form of different
container classes. Existing vfio container is named legacy container
(equivalent with legacy iommu backend in this series), while the new
iommufd based container is named as iommufd container (may also mentioned
as iommufd backend in this series). The two backend types have their own
way to setup secure context and dma management interface. Below diagram
shows how it looks like with both BEs.

                    VFIO                            AddressSpace/Memory
    +--------+   +-----------+   +-------+
    |  pci   |   |  platform |   |   ap  | ...
    +---+----+   +-----+-----+   +---+---+        +----------------------+
        |              |             |            |   AddressSpace       |
       \|/            \|/           \|/           +------------+---------+
    +---+--------------+-------------+--------+               /
    |           VFIOAddressSpace              | <------------+
    |                  |                      |  MemoryListener
    |          VFIOContainer list             |
    +-------+----------------------------+----+
            |                            |
           \|/                          \|/
    +-------+------+            +--------+----------+
    |   iommufd    |            |    vfio legacy    |
    |  container   |            |     container     |
    +-------+------+            +--------+----------+
            |                            |
            | /dev/iommu                 | /dev/vfio/vfio
            | /dev/vfio/devices/vfioX    | /dev/vfio/$group_id
 Userspace  |                            |
 ===========+============================+================================
 Kernel     |  device fd                 |
            +---------------+            | group/container fd
            | (BIND_IOMMUFD |            | (SET_CONTAINER/SET_IOMMU)
            |  ATTACH_IOAS) |            | device fd
            |              \|/          \|/
            |       +-------+------------+-----------------+
    iommufd |       |                vfio                  |
(map/unmap  |       +-------+--------------------+---------+
 ioas_copy) |               |                    | map/unmap
           \|/             \|/                  \|/
     +------+------+  +-----+------+      +------+--------+
     | iommfd core |  |  device    |      |  vfio iommu   |
     +-------------+  +------------+      +---------------+

[Secure Context setup]
- iommufd BE: uses device fd and iommufd to setup secure context
              (bind_iommufd, attach_ioas)
- vfio legacy BE: uses group fd and container fd to setup secure context
                  (set_container, set_iommu)
[Device access]
- iommufd BE: device fd is opened through /dev/vfio/devices/vfioX
- vfio legacy BE: device fd is retrieved from group fd ioctl
[DMA Mapping flow]
- VFIOAddressSpace receives MemoryRegion add/del via MemoryListener
- VFIO populates DMA map/unmap via the container BEs
  *) iommufd BE: uses iommufd
  *) vfio legacy BE: uses container fd

This series qomifies the VFIOContainer object which acts as a base class
for a container. This base class is derived into the legacy VFIO container
and the new iommufd based container. The base class implements generic code
such as code related to memory_listener and address space management whereas
the dervied class implements callbacks that depend on the kernel user space
being used.

The selection of the backend is made on a device basis using the new
iommufd option (on/off/auto). By default the iommufd backend is selected
if supported by the host and by QEMU (iommufd KConfig). This option is
currently available only for the vfio-pci device. For other types of
devices, it does not yet exist and the legacy BE is chosen by default.

Test done:
- PCI and Platform device were tested
- ccw and ap were only compile-tested.
- limited hotplug test
- vIOMMU test run for both legacy and iommufd backends

This series was co-developped by Eric Auger and me based on the exploration
iommufd kernel[2], complete code of this series is available in[3]. As
iommufd kernel is in the early step (only iommufd generic interface is in
mailing list), so this series hasn't made the iommufd backend fully on par
with legacy backend w.r.t. features like live migration, kvm group add/del
and etc.

TODOs:
- Add DMA alias check for iommufd BE (group level)
- Cleanup the VFIODevice fields as it's used in both BEs
- Add locks
- More tests
- ....

Patch Overview:

- Preparation works:
0001-scripts-update-linux-headers-Add-iommufd.h.patch
0002-linux-headers-Import-latest-vfio.h-and-iommufd.h.patch
0003-vfio-pci-Use-vbasedev-local-variable-in-vfio_realize.patch
0004-vfio-common-Rename-VFIOGuestIOMMU-iommu-into-iommu_m.patch
0005-vfio-common-Split-common.c-into-common.c-container.c.patch

- Introduce container base object and refactor the old code to be legacy backend
0006-vfio-Add-base-object-for-VFIOContainer.patch
0007-vfio-container-Introduce-vfio_attach-detach_device.patch
0008-vfio-platform-Use-vfio_-attach-detach-_device.patch
0009-vfio-ap-Use-vfio_-attach-detach-_device.patch
0010-vfio-ccw-Use-vfio_-attach-detach-_device.patch
0011-vfio-container-obj-Introduce-attach-detach-_device-c.patch

- Add iommufd support
0012-hw-iommufd-Creation.patch
0013-vfio-iommufd-Implement-iommufd-backend.patch
0014-vfio-iommufd-Add-IOAS_COPY_DMA-support.patch

- Add backend selection for vfio pci
0015-vfio-as-Allow-the-selection-of-a-given-iommu-backend.patch
0016-vfio-pci-Add-an-iommufd-option.patch

[1] https://lore.kernel.org/kvm/0-v1-e79cd8d168e8+6-iommufd_jgg@nvidia.com/
[2] https://github.com/luxis1999/iommufd/tree/iommufd-v5.17-rc6
[3] https://github.com/luxis1999/qemu/tree/qemu-for-5.17-rc6-vm-rfcv1

Base commit: 4bf58c7213 virtio-iommu: use-after-free fix

Signed-off-by: Yi Liu <yi.l.liu@intel.com>